refactor: Renamed sfx folder to audio.
