+++ src/audio/musicplayer.js
